Added clean architecture with mvp
