--- app/src/main/java/ru/androidschool/intensiv/data/Actor.kt
@@ -1,6 +1,0 @@
-package ru.androidschool.intensiv.data
-
-data class Actor(
-    val name: String,
-    val img: Int
-)
