--- app/src/main/java/ru/androidschool/intensiv/data/Dates.kt
@@ -1,10 +1,0 @@
-package ru.androidschool.intensiv.data
-
-import com.google.gson.annotations.SerializedName
-
-data class Dates(
-    @SerializedName("maximum")
-    val maximum: String?,
-    @SerializedName("minimum")
-    val minimum: String?
-)
