--- app/src/main/java/ru/androidschool/intensiv/data/DetailedMovie.kt
@@ -1,6 +1,0 @@
-package ru.androidschool.intensiv.data
-
-data class DetailedMovie(
-    val credits: MovieCredits?,
-    val info: MovieInfo?
-)
