--- app/src/main/java/ru/androidschool/intensiv/data/MovieParameter.kt
@@ -1,6 +1,0 @@
-package ru.androidschool.intensiv.data
-
-data class MovieParameter(
-    val title: String,
-    val value: String
-)
