+++ app/src/main/java/ru/androidschool/intensiv/data/dto/Actor.kt
@@ -1,0 +1,6 @@
+package ru.androidschool.intensiv.data.dto
+
+data class Actor(
+    val name: String,
+    val img: Int
+)
