+++ app/src/main/java/ru/androidschool/intensiv/data/dto/Dates.kt
@@ -1,0 +1,10 @@
+package ru.androidschool.intensiv.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class Dates(
+    @SerializedName("maximum")
+    val maximum: String?,
+    @SerializedName("minimum")
+    val minimum: String?
+)
