+++ app/src/main/java/ru/androidschool/intensiv/data/dto/DetailedMovie.kt
@@ -1,0 +1,6 @@
+package ru.androidschool.intensiv.data.dto
+
+data class DetailedMovie(
+    val credits: MovieCredits?,
+    val info: MovieInfo?
+)
