+++ app/src/main/java/ru/androidschool/intensiv/data/dto/MovieParameter.kt
@@ -1,0 +1,6 @@
+package ru.androidschool.intensiv.data.dto
+
+data class MovieParameter(
+    val title: String,
+    val value: String
+)
