+++ app/src/main/java/ru/androidschool/intensiv/domain/repository/TvShowRepository.kt
@@ -1,0 +1,8 @@
+package ru.androidschool.intensiv.domain.repository
+
+import io.reactivex.Single
+import ru.androidschool.intensiv.data.dto.MovieResponse
+
+interface TvShowRepository {
+    fun getTvShow(): Single<MovieResponse>
+}
