--- app/src/main/java/ru/androidschool/intensiv/network/MovieApiClient.kt
@@ -1,28 +1,0 @@
-package ru.androidschool.intensiv.network
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-import ru.androidschool.intensiv.BuildConfig
-
-object MovieApiClient {
-
-    private val client: OkHttpClient = OkHttpClient.Builder()
-        .addInterceptor(HttpLoggingInterceptor().apply {
-            this.level = HttpLoggingInterceptor.Level.BODY
-        })
-        .build()
-
-    val apiClient: MovieApiInterface by lazy {
-        val retrofit = Retrofit.Builder()
-            .baseUrl(BuildConfig.BASE_URL)
-            .client(client)
-            .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-        return@lazy retrofit.create(MovieApiInterface::class.java)
-    }
-}
