+++ app/src/main/java/ru/androidschool/intensiv/room/WishListMoviesDB.kt
@@ -2,7 +2,7 @@ package ru.androidschool.intensiv.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import ru.androidschool.intensiv.data.DetailedMovieEntity
+import ru.androidschool.intensiv.data.dto.DetailedMovieEntity
 
 @Database(entities = [DetailedMovieEntity::class], version = 1)
 abstract class AppDB : RoomDatabase() {
