+++ app/src/main/java/ru/androidschool/intensiv/utils.kt
@@ -7,18 +7,17 @@ import io.reactivex.Observable
 import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import kotlinx.android.synthetic.main.actor_item.*
-import ru.androidschool.intensiv.network.MovieApiClient
+import ru.androidschool.intensiv.data.network.MovieApiClient
 
 // QUESTION: такие extensions? подавлять checkResult?
 
 @SuppressLint("CheckResult")
-fun <T> Observable<T>.myObserve(): Observable<T> =
+fun <T> Observable<T>.applySchedulers(): Observable<T> =
     this.subscribeOn(Schedulers.io())
         .observeOn(AndroidSchedulers.mainThread())
 
 @SuppressLint("CheckResult")
-fun <T> Single<T>.myObserve(): Single<T> =
+fun <T> Single<T>.applySchedulers(): Single<T> =
     this.subscribeOn(Schedulers.io())
         .observeOn(AndroidSchedulers.mainThread())
